function that prints integers

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
 
 void print_number(int n)
 {
-	double len, res, i, temp, expo;
+	float len, res, i, temp, expo;
 
 	res = n;
 	expo = len =  1;
